backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
boto3==1.35.24
google-cloud-storage==2.18.2
celery==5.4.0
redis==5.0.8
python-jose==3.3.0
httpx==0.27.2
PyMuPDF>=1.24.0
pix2tex>=0.1.1
Pillow>=10.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-pptx==1.0.2
langdetect==1.0.9
thefuzz[speedup]==0.22.1
